# Conflicts:
#	app/src/main/java/com/madrid/movio/di/DataModule.kt
#	data/src/main/java/com/madrid/data/dataSource/datastore/UserPreferencesImpl.kt
#	data/src/main/java/com/madrid/data/dataSource/remote/MovioApi.kt
#	data/src/main/java/com/madrid/data/dataSource/remote/RemoteDataSourceImpl.kt
#	data/src/main/java/com/madrid/data/repositories/SearchRepositoryImpl.kt
#	presentation/src/main/java/com/madrid/presentation/screens/detailsScreen/seriesDetails/SeriesDetailsScreen.kt
#	presentation/src/main/java/com/madrid/presentation/viewModel/detailsViewModel/TopCastViewModel.kt

+++ presentation/src/main/java/com/madrid/presentation/viewModel/detailsViewModel/ReviewsScreenViewModel.kt
@@ -30,7 +30,7 @@ class ReviewsScreenViewModel @Inject constructor(
 
     private fun loadMovieReviews() {
         tryToExecute(
-            function = { getMovieReviewsUseCase(args.mediaId.toInt()) },
+            function = { getMovieReviewsUseCase(args.mediaId) },
             onSuccess = { reviews ->
                 updateState {
                     it.copy(reviews = reviews.map { review ->
@@ -44,7 +44,7 @@ class ReviewsScreenViewModel @Inject constructor(
 
     private fun loadSeriesReviews() {
         tryToExecute(
-            function = { getSeriesReviewsUseCase(args.mediaId.toInt()) },
+            function = { getSeriesReviewsUseCase(args.mediaId) },
             onSuccess = { reviews ->
                 updateState {
                     it.copy(reviews = reviews.map { review ->
